+++ doker/rst2pdf/svgimage.py
@@ -1,52 +1,45 @@
 # -*- coding: utf-8 -*-
 # See LICENSE.txt for licensing terms
 
-import os
-
+from reportlab.lib.enums import TA_CENTER, TA_LEFT, TA_RIGHT
 from reportlab.platypus import Flowable, Paragraph
-from reportlab.lib.enums import TA_LEFT, TA_CENTER, TA_RIGHT
 
 from .log import log
-from .opt_imports import LazyImports
 
+from svglib.svglib import svg2rlg
 
 class SVGImage(Flowable):
 
-    @classmethod
-    def available(self):
-        if LazyImports.svg2rlg:
-            return True
-        return False
-
-    def __init__(self, filename, width=None, height=None, kind='direct',
-                                     mask=None, lazy=True, srcinfo=None):
+    def __init__(self,
+                 filename,
+                 width=None,
+                 height=None,
+                 kind='direct',
+                 mask=None,
+                 lazy=True,
+                 srcinfo=None):
         Flowable.__init__(self)
-        ext = os.path.splitext(filename)[-1]
         self._kind = kind
-        # Prefer svg2rlg for SVG, as it works better
-        if LazyImports.svg2rlg:
-            self._mode = 'svg2rlg'
-            self.doc = LazyImports.svg2rlg.svg2rlg(filename)
-            self.imageWidth = width
-            self.imageHeight = height
-            x1, y1, x2, y2 = self.doc.getBounds()
-            # Actually, svg2rlg's getBounds seems broken.
-            self._w, self._h = x2, y2
-            if not self.imageWidth:
-                self.imageWidth = self._w
-            if not self.imageHeight:
-                self.imageHeight = self._h
-        else:
-            self._mode = None
-            log.error("SVG support not enabled,"
-                " please install svg2rlg.")
+        self._mode = 'svg2rlg'
+        self.doc = svg2rlg(filename)
+        self.imageWidth = width
+        self.imageHeight = height
+        x1, y1, x2, y2 = self.doc.getBounds()
+        # Actually, svg2rlg's getBounds seems broken.
+        self._w, self._h = x2, y2
+        if not self.imageWidth:
+            self.imageWidth = self._w
+        if not self.imageHeight:
+            self.imageHeight = self._h
         self.__ratio = float(self.imageWidth)/self.imageHeight
-        if kind in ['direct','absolute']:
+        if kind in ['direct', 'absolute']:
             self.drawWidth = width or self.imageWidth
             self.drawHeight = height or self.imageHeight
-        elif kind in ['bound','proportional']:
-            factor = min(float(width)/self.imageWidth,float(height)/self.imageHeight)
-            self.drawWidth = self.imageWidth*factor
+        elif kind in ['bound', 'proportional']:
+            factor = min(
+                float(width) / self.imageWidth,
+                float(height) / self.imageHeight)
+            self.drawWidth = self.imageWidth * factor
             self.drawHeight = self.imageHeight*factor
 
     def wrap(self, aW, aH):